requirements.txt
ffpyplayer==4.4.0
numpy==1.24.2
opencv-python==4.7.0.72
Pillow==9.4.0
PyAudio==0.2.13
